WIP extract AI prompt input builders
