+++ applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Application/AI/Operations/ApplicationAnalysisService.cs
@@ -1,12 +1,8 @@
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Linq;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using Unity.AI.Models;
 using Unity.AI.Prompts;
 using Unity.AI.Requests;
 using Unity.AI.Runtime;
@@ -24,12 +20,6 @@ namespace Unity.AI.Operations
         IAIGenerationPrerequisiteValidator aiGenerationPrerequisiteValidator,
         ILogger<ApplicationAnalysisService> logger) : IApplicationAnalysisService, ITransientDependency
     {
-        private const string ComponentsKey = "components";
-        private static readonly HashSet<string> ExcludedSchemaKeys = new(StringComparer.OrdinalIgnoreCase)
-        {
-            "applicantAgent"
-        };
-
         public async Task<string> RegenerateAndSaveAsync(Guid applicationId, string? promptVersion = null, CancellationToken cancellationToken = default)
         {
             await aiGenerationPrerequisiteValidator.EnsureApplicationAnalysisAvailableAsync(applicationId);
@@ -39,20 +29,11 @@ namespace Unity.AI.Operations
             var attachments = await applicationChefsFileAttachmentRepository.GetListAsync(a => a.ApplicationId == applicationId);
             var formSchema = await GetFormSchemaAsync(formSubmission?.ApplicationFormVersionId);
 
-            var attachmentSummaries = attachments
-                .Where(a => !string.IsNullOrWhiteSpace(a.AISummary))
-                .Select(a => new AIAttachmentItem
-                {
-                    Name = string.IsNullOrWhiteSpace(a.FileName) ? "attachment" : a.FileName.Trim(),
-                    Summary = a.AISummary!.Trim()
-                })
-                .ToList();
-
-            object formFieldConfiguration = new { message = "Form configuration not available." };
-            if (formSubmission?.ApplicationFormVersionId != null)
-            {
-                formFieldConfiguration = await ExtractFormFieldConfigurationAsync(formSubmission.ApplicationFormVersionId.Value);
-            }
+            var attachmentSummaries = PromptDataPayloadBuilder.BuildAttachmentSummaries(attachments);
+            var formFieldConfiguration = await PromptDataPayloadBuilder.BuildFormFieldConfigurationAsync(
+                applicationFormVersionRepository,
+                formSubmission?.ApplicationFormVersionId,
+                logger);
 
             var analysis = await aiService.GenerateApplicationAnalysisAsync(new ApplicationAnalysisRequest
             {
@@ -87,115 +68,5 @@ namespace Unity.AI.Operations
             }
         }
 
-        private async Task<object> ExtractFormFieldConfigurationAsync(Guid formVersionId)
-        {
-            try
-            {
-                var formVersion = await applicationFormVersionRepository.GetAsync(formVersionId);
-                if (formVersion == null || string.IsNullOrEmpty(formVersion.FormSchema))
-                {
-                    return new { message = "Form configuration not available." };
-                }
-
-                var schema = JObject.Parse(formVersion.FormSchema);
-                var components = schema[ComponentsKey] as JArray;
-                if (components == null || components.Count == 0)
-                {
-                    return new { message = "No form fields configured." };
-                }
-
-                var requiredFields = new List<string>();
-                var optionalFields = new List<string>();
-                ExtractFieldRequirements(components, requiredFields, optionalFields, string.Empty);
-
-                return new
-                {
-                    required_fields = requiredFields,
-                    optional_fields = optionalFields
-                };
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex, "Error extracting form field configuration for form version {FormVersionId}", formVersionId);
-                return new { message = "Form configuration could not be extracted." };
-            }
-        }
-
-        private static void ExtractFieldRequirements(JArray components, List<string> requiredFields, List<string> optionalFields, string currentPath)
-        {
-            foreach (var component in components.OfType<JObject>())
-            {
-                var key = component["key"]?.ToString();
-                var label = component["label"]?.ToString();
-                var type = component["type"]?.ToString();
-                var skipTypes = new HashSet<string> { "button", "simplebuttonadvanced", "html", "htmlelement", "content", "simpleseparator" };
-
-                if (string.IsNullOrEmpty(key) || string.IsNullOrEmpty(type) || skipTypes.Contains(type) || ExcludedSchemaKeys.Contains(key))
-                {
-                    ProcessNestedFieldRequirements(component, type, requiredFields, optionalFields, currentPath);
-                    continue;
-                }
-
-                var displayName = !string.IsNullOrEmpty(label) ? $"{label} ({key})" : key;
-                var fullPath = string.IsNullOrEmpty(currentPath) ? displayName : $"{currentPath} > {displayName}";
-                var validate = component["validate"] as JObject;
-                var isRequired = validate?["required"]?.Value<bool>() ?? false;
-
-                if (component["input"]?.Value<bool>() == true)
-                {
-                    if (isRequired) requiredFields.Add(fullPath);
-                    else optionalFields.Add(fullPath);
-                }
-
-                ProcessNestedFieldRequirements(component, type, requiredFields, optionalFields, fullPath);
-            }
-        }
-
-        private static void ProcessNestedFieldRequirements(JObject component, string? type, List<string> requiredFields, List<string> optionalFields, string currentPath)
-        {
-            switch (type)
-            {
-                case "panel":
-                case "simplepanel":
-                case "fieldset":
-                case "well":
-                case "container":
-                case "datagrid":
-                case "table":
-                    if (component[ComponentsKey] is JArray nestedComponents)
-                    {
-                        ExtractFieldRequirements(nestedComponents, requiredFields, optionalFields, currentPath);
-                    }
-                    break;
-                case "columns":
-                case "simplecols2":
-                case "simplecols3":
-                case "simplecols4":
-                    if (component["columns"] is JArray columns)
-                    {
-                        foreach (var column in columns.OfType<JObject>())
-                        {
-                            if (column[ComponentsKey] is JArray columnComponents)
-                            {
-                                ExtractFieldRequirements(columnComponents, requiredFields, optionalFields, currentPath);
-                            }
-                        }
-                    }
-                    break;
-                case "tabs":
-                case "simpletabs":
-                    if (component[ComponentsKey] is JArray tabs)
-                    {
-                        foreach (var tab in tabs.OfType<JObject>())
-                        {
-                            if (tab[ComponentsKey] is JArray tabComponents)
-                            {
-                                ExtractFieldRequirements(tabComponents, requiredFields, optionalFields, currentPath);
-                            }
-                        }
-                    }
-                    break;
-            }
-        }
     }
 }
